TIP-842: Move attribute option subscriber test to Structure
