+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/services.yml
@@ -8,3 +8,11 @@ services:
             - '%pim_catalog.entity.group_type.class%'
         tags:
             - { name: form.type, alias: pim_enrich_grouptype }
+
+    pim_catalog.event_subscriber.attribute_option_removal:
+        class: 'Akeneo\Pim\Structure\Bundle\EventListener\AttributeOptionRemovalSubscriber'
+        arguments:
+            - '@pim_catalog.doctrine.query.find_family_variants_identifiers_by_attribute_axes'
+            - '@pim_catalog.query.product_and_product_model_query_builder_factory'
+        tags:
+            - { name: kernel.event_subscriber }
